Major Logic changes
The idea is to have real time input verification for the
`.answer` box. And to show warning in real time in `.history`.
Example: User entered "1.111111...." but screen only showed "1.1"
and in `.history` following warning would come "only 1 decimal
allowed" and "only 1 decimal point allowed".

Previously, there were two functions for handling clicks on
`.number` and `.symbol`. But, the code was becoming very hard
to read. So, now, try to use regex to solve the problem.