Handle unary and binary operations
numberHandler: handles numbers and decimal as input. If the input
is a decimal point, throws error on secondary display as
decimal is automatically added. This also adds the number in an
calculator object and displays the object in main display

symbolHandler: handles symbols. currently, one binary and one unary
symbol (+) and (square root) is handled.