Add fix for input, support C & AC, scrollbar
Fixed input, it was not possible to enter more than 1 digit in left
of decimal digit. This was happening because after calculating one
line of operation, we were storing the answer in the calculator.a
but we used a flawed regex. Fixed the regex to accept negative values
and decimal points.

Added support for C - clears ANSWER screen and resets calculator object.
AC clears HISTORY and ANSWER and resets calculator object.

Added overflow in ANSWER so that overflow does not cause leaking text.